requirements/tests.txt
pytest==5.4.3
pytest-cache==1.0
pytest-django==3.9.0
pytest-sugar==0.9.4
pytest-watch==4.2.0
pytest-mock==3.2.0
pytest-cov==2.10.0
requests-mock==1.8.0
coverage
